+++ src/differential_privacy/budget.rs
@@ -57,6 +57,3 @@ impl Hash for Budget {
 }
 
 impl Eq for Budget {}
-
-
-