+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Differential privacy in rules [MR156](https://github.com/Qrlew/qrlew/pull/156)
 ### Fixed
 - differential privacy for the new formalism [MR155](https://github.com/Qrlew/qrlew/pull/155)
+### Added
+- `rewriting_rul.RewriteVisitor` outputs a `RelationWithPrivateQuery` [MR157](https://github.com/Qrlew/qrlew/pull/157)
 
 
 ## [0.4.0] - 2023-10-19